# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: zTech CRM API
  description: API del CRM de salones de eventos corporativos — UNLaM GADS1.
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/contacts/{id}:
    get:
      tags:
      - Contacts
      summary: Detalle de un contacto
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactDetailResponse"
        "404":
          description: No existe un contacto con ese id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactDetailResponse"
    put:
      tags:
      - Contacts
      summary: Edita un contacto
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
        required: true
      responses:
        "200":
          description: Contacto actualizado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: "No existe un contacto con ese id, o la empresa relacionada\
            \ no existe"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "409":
          description: Ya existe un contacto con ese documento
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /api/v1/companies/{id}:
    get:
      tags:
      - Companies
      summary: "Detalle de una empresa, con sus contactos (BE-CUS-03)"
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyDetailResponse"
        "404":
          description: No existe una empresa con ese id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyDetailResponse"
    put:
      tags:
      - Companies
      summary: Edita una empresa
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRequest"
        required: true
      responses:
        "200":
          description: Empresa actualizada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "404":
          description: No existe una empresa con ese id
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "409":
          description: Ya existe una empresa con ese CUIT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
...
  /api/v1/contacts:
    get:
      tags:
      - Contacts
      summary: "Lista contactos del tenant, paginado"
      operationId: list_1
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseContactResponse"
    post:
      tags:
      - Contacts
      summary: Crea un contacto
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
        required: true
      responses:
        "201":
          description: Contacto creado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: La empresa relacionada no existe
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "409":
          description: Ya existe un contacto con ese documento
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /api/v1/companies:
    get:
      tags:
      - Companies
      summary: "Lista empresas del tenant, paginado"
      operationId: list_2
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseCompanyResponse"
    post:
      tags:
      - Companies
      summary: Crea una empresa
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRequest"
        required: true
      responses:
        "201":
          description: Empresa creada
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "400":
          description: Datos inválidos
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
        "409":
          description: Ya existe una empresa con ese CUIT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompanyResponse"
...
components:
  schemas:
...
    ContactRequest:
      type: object
      properties:
        companyId:
          type: integer
          format: int64
          description: Id de la empresa relacionada. Vacío si es cliente individual
            (Módulo 1 de la consigna).
        firstName:
          type: string
          example: María
          minLength: 1
        lastName:
          type: string
          example: Gómez
          minLength: 1
        document:
          type: string
          description: Opcional
          example: "30123456"
        position:
          type: string
          description: "Cargo, si corresponde"
          example: Responsable de compras
        email:
          type: string
          format: email
          example: maria.gomez@empresa.com
        phone:
          type: string
          example: +54 11 4000-0000
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
          description: "Id del usuario responsable comercial, si ya está asignado"
        originId:
          type: integer
          format: int64
          description: Id del origen comercial
        notes:
          type: string
      required:
      - firstName
      - lastName
      - status
    ContactResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
    CompanyRequest:
      type: object
      properties:
        name:
          type: string
          example: Eventos del Sur SA
          minLength: 1
        cuit:
          type: string
          description: "Opcional. Formato: 11 dígitos, con o sin guiones."
          example: 30-71234567-9
        industry:
          type: string
          example: Turismo y eventos
        email:
          type: string
          format: email
          example: contacto@eventosdelsur.com
        phone:
          type: string
          example: +54 11 4000-0000
        address:
          type: string
          example: "Av. Rivadavia 4500, San Justo"
        website:
          type: string
          example: https://eventosdelsur.com
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
          description: "Id del usuario responsable comercial, si ya está asignado"
        originId:
          type: integer
          format: int64
          description: Id del origen comercial
        notes:
          type: string
      required:
      - name
      - status
    CompanyResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cuit:
          type: string
        industry:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    CompanySummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    ContactSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
...
    PageResponseContactResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ContactResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    ContactDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        company:
          $ref: "#/components/schemas/CompanySummaryResponse"
        firstName:
          type: string
        lastName:
          type: string
        document:
          type: string
        position:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
        originId:
          type: integer
          format: int64
        notes:
          type: string
    PageResponseCompanyResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/CompanyResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    CompanyDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cuit:
          type: string
        industry:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        website:
          type: string
        status:
          type: string
          enum:
          - POTENCIAL
          - CLIENTE
          - INACTIVO
          - NO_CONTACTAR
        salesRepId:
          type: integer
          format: int64
        originId:
          type: integer
          format: int64
        notes:
          type: string
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/ContactSummaryResponse"
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT